rcon/requirements.txt
boto3==1.35.71
botocore==1.35.71
factorio-rcon-py==2.1.3
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
